moved ColorMapPicker.svelte
