+++ src/components/controls/ColorMapPicker.svelte
